Make buckets and users synchronous with folders

+++ src/controllers/BucketController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express'
 import { getRepository, Repository, getManager } from 'typeorm'
+import { getEnvFolder } from '../main'
 import FileManager from '../services/filemanager'
 import Authentifier from '../services/authentifier'
 import Bucket from '../entity/Bucket'
@@ -34,11 +35,8 @@ class BucketController {
         .then(
           (result): Response => {
             // ~/myS3DATA/$USER_UUID/$BUCKET_NAME/$BLOB_NAME
-            const filemanager = new FileManager('~/myS3DATA')
-            console.log(
-              filemanager.createFolder(`/${result.user.id}/${result.name}/`),
-            )
-            //console.log(result)
+            // Create folder with bucket name
+            getEnvFolder.createFolder(`${result.user.id}/${result.name}`)
 
             return res.send(result)
           },
@@ -47,7 +45,9 @@ class BucketController {
           res.status(400).send(error)
         })
     } else {
-      return res.status(400).send({message: 'ERROR : Missing Bearer token in your Authorizations'})
+      return res.status(400).send({
+        message: 'ERROR : Missing Bearer token in your Authorizations',
+      })
     }
   }
 
@@ -71,29 +71,51 @@ class BucketController {
           .status(400)
           .send({ message: "Bucket doesn't exists in database" })
       }
+      const oldName = bucket.name
       bucketRepository.merge(bucket, req.body)
-      await bucketRepository.save(bucket).then(
-        (result): Response => {
-          return res.send(result)
-        },
-      )
+      await bucketRepository.save(bucket).then((result): Response => {
+        // Rename folder with new bucket name
+        getEnvFolder.renameFolder(
+          `${authUser.user.id}/${oldName}`,
+          `${authUser.user.id}/${bucket.name}`,
+        )
+        return res.send(result)
+      })
     } else {
-      return res
-        .status(400)
-        .send({
-          message: 'ERROR : Missing Bearer token in your Authorizations',
-        })
+      return res.status(400).send({
+        message: 'ERROR : Missing Bearer token in your Authorizations',
+      })
     }
   }
 
   // Delete bucket
   static deleteBucket = async (req: Request, res: Response): Promise<void> => {
-    const bucketRepository: Repository<Bucket> = getRepository(Bucket)
-    await bucketRepository.delete(req.params.id).then(
-      (result): Response => {
-        return res.send(result)
-      },
-    )
+    if (req.headers.authorization) {
+      const userToken = req.headers.authorization.replace('Bearer ', '')
+      const auth = new Authentifier(userToken)
+      const authUser = await auth.getUser()
+      if (!authUser.user) return res.status(400).send(authUser.message)
+
+      const bucketRepository: Repository<Bucket> = getRepository(Bucket)
+      const bucket = await bucketRepository.findOne({
+        where: { id: req.params.id },
+      })
+      if (bucket === undefined) {
+        return res
+          .status(400)
+          .send({ message: "Bucket doesn't exists in database" })
+      }
+
+      await bucketRepository.delete(req.params.id).then(
+        (result): Response => {
+          getEnvFolder.deleteFolder(`${authUser.user.id}/${bucket.name}`)
+          return res.send(result)
+        })
+    } else {
+      return res.status(400).send({
+        message: 'ERROR : Missing Bearer token in your Authorizations',
+      })
+    }
   }
 }
